feat: merge snippets to single source
