+++ docs/configuration/reference.md
@@ -435,6 +435,7 @@ sources.providers = {
     score_offset = 0, -- Boost/penalize the score of the items
     override = nil, -- Override the source's functions
   },
+
   path = {
     name = 'Path',
     module = 'blink.cmp.sources.path',
@@ -447,9 +448,12 @@ sources.providers = {
       show_hidden_files_by_default = false,
     }
   },
+
   snippets = {
     name = 'Snippets',
     module = 'blink.cmp.sources.snippets',
+
+    -- For `snippets.preset == 'default'`
     opts = {
       friendly_snippets = true,
       search_paths = { vim.fn.stdpath('config') .. '/snippets' },
@@ -462,17 +466,22 @@ sources.providers = {
       -- Set to '+' to use the system clipboard, or '"' to use the unnamed register
       clipboard_register = nil,
     }
-  },
-  luasnip = {
-    name = 'Luasnip',
-    module = 'blink.cmp.sources.luasnip',
+
+    -- For `snippets.preset == 'luasnip'`
     opts = {
       -- Whether to use show_condition for filtering snippets
       use_show_condition = true,
       -- Whether to show autosnippets in the completion list
       show_autosnippets = true,
     }
+
+    -- For `snippets.preset == 'mini_snippets'`
+    opts = {
+      -- Whether to use a cache for completion items
+      use_items_cache = true,
+    }
   },
+
   buffer = {
     name = 'Buffer',
     module = 'blink.cmp.sources.buffer',
