+++ lua/blink/cmp/sources/snippets/default/scan.lua
